+++ LogFileManager.cpp
@@ -2,7 +2,9 @@
 #include "formatString.h"
 #include "log.h"
 
-#include <stdlib.h>
+#include <chrono>
+#include <iomanip>
+
 #include <boost/filesystem.hpp>
 #include <boost/system/error_code.hpp>
 
@@ -22,41 +24,32 @@ LogFileManager* LogFileManager::instance()
 
 LogFileManager::LogFileManager()
 {
-    _homeDir    = std::string(getenv("HOME"));
-    _logDir     = formatString("%s/%s_1", _homeDir.c_str(), _logDirPrefix);
+    _homeDir = std::string(getenv("HOME"));
 }
 
 void LogFileManager::detectorsStarted()
 {
-    if (_detectorStartIndex++ == 0) {
-        bs::error_code errorCode;
-        
-        for (int i=10; i>=1; i--) {
-            auto oldName = formatString("%s/%s_%d", _homeDir.c_str(), _logDirPrefix, i);
-            auto newName = formatString("%s/%s_%d", _homeDir.c_str(), _logDirPrefix, i + 1);
+    auto now_time_t = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+    auto now_utc    = *std::gmtime(&now_time_t);
 
-            if (bf::exists(newName, errorCode)) {
-                logDebug() << "Removing " << newName;
-                bf::remove_all(newName, errorCode);
-                if (errorCode) {
-                    logDebug() << "Failed to remove " << newName << ": " << errorCode.message();
-                }
-            }
-            
-            bf::rename(oldName.c_str(), newName.c_str(), errorCode);
-            if (errorCode) {
-                logDebug() << "Failed to rename " << oldName << " to " << newName << ": " << errorCode.message();
-            }
-        }
+    char buffer[80];
+    std::strftime(buffer, sizeof(buffer), "%Y-%m-%d_%H-%M", &now_utc);
 
-        bf::create_directory(_logDir.c_str(), errorCode);
-        if (errorCode) {
-            logDebug() << "Failed to create directory " << _logDir << ": " << errorCode.message();
-        }
+    _logDir = formatString("%s-%s", _logDirPrefix, buffer);
+
+    bs::error_code errorCode;
+    bf::create_directory(_logDir.c_str(), errorCode);
+    if (errorCode) {
+        logDebug() << "Failed to create directory " << _logDir << ": " << errorCode.message();
     }
+}
+
+void LogFileManager::detectorsStopped()
+{
+
 }
 
 std::string LogFileManager::filename(const char* root, const char* extension)
 {
-    return formatString("%s/%s.%d.%s", _logDir.c_str(), root, _detectorStartIndex, extension);
+    return formatString("%s/%s.%s", _logDir.c_str(), root, extension);
 }
