+++ log.h
@@ -25,11 +25,12 @@ void set_color(LogColor LogColor);
 
 class LogDetailed {
 public:
-    LogDetailed(const char* filename, int filenumber) :
-        _s(),
-        _caller_filename(filename),
-        _caller_filenumber(filenumber)
-    {}
+    LogDetailed(const char* filename, int filenumber);
+    LogDetailed(const LogDetailed&) = delete;
+
+    virtual ~LogDetailed();
+
+    void operator=(const LogDetailed&) = delete;
 
     LogDetailed& operator<<(uint8_t& x)
     {
@@ -51,63 +52,6 @@ public:
         return *this;
     }
 
-    virtual ~LogDetailed()
-    {
-
-        switch (_log_level) {
-            case LogLevel::Debug:
-                set_color(LogColor::Green);
-                break;
-            case LogLevel::Info:
-                set_color(LogColor::Blue);
-                break;
-            case LogLevel::Warn:
-                set_color(LogColor::Yellow);
-                break;
-            case LogLevel::Err:
-                set_color(LogColor::Red);
-                break;
-        }
-
-        _logMutex.lock();
-
-        // Time output taken from:
-        // https://stackoverflow.com/questions/16357999#answer-16358264
-        time_t rawtime;
-        time(&rawtime);
-        struct tm* timeinfo = localtime(&rawtime);
-        char time_buffer[10]{}; // We need 8 characters + \0
-        strftime(time_buffer, sizeof(time_buffer), "%I:%M:%S", timeinfo);
-        std::cout << "[" << time_buffer;
-
-        switch (_log_level) {
-            case LogLevel::Debug:
-                std::cout << "|D] ";
-                break;
-            case LogLevel::Info:
-                std::cout << "|I] ";
-                break;
-            case LogLevel::Warn:
-                std::cout << "|W] ";
-                break;
-            case LogLevel::Err:
-                std::cout << "|E] ";
-                break;
-        }
-
-        set_color(LogColor::Reset);
-
-        std::cout << _s.str();
-        std::cout << " (" << _caller_filename << ":" << std::dec << _caller_filenumber << ")";
-
-        std::cout << std::endl;
-
-        _logMutex.unlock();
-    }
-
-    LogDetailed(const LogDetailed&) = delete;
-    void operator=(const LogDetailed&) = delete;
-
 protected:
     LogLevel _log_level = LogLevel::Debug;
 
@@ -121,7 +65,8 @@ private:
 
 class LogDebugDetailed : public LogDetailed {
 public:
-    LogDebugDetailed(const char* filename, int filenumber) : LogDetailed(filename, filenumber)
+    LogDebugDetailed(const char* filename, int filenumber) 
+        : LogDetailed(filename, filenumber)
     {
         _log_level = LogLevel::Debug;
     }
@@ -129,7 +74,8 @@ public:
 
 class LogInfoDetailed : public LogDetailed {
 public:
-    LogInfoDetailed(const char* filename, int filenumber) : LogDetailed(filename, filenumber)
+    LogInfoDetailed(const char* filename, int filenumber) 
+        : LogDetailed(filename, filenumber)
     {
         _log_level = LogLevel::Info;
     }
@@ -137,7 +83,8 @@ public:
 
 class LogWarnDetailed : public LogDetailed {
 public:
-    LogWarnDetailed(const char* filename, int filenumber) : LogDetailed(filename, filenumber)
+    LogWarnDetailed(const char* filename, int filenumber) 
+        : LogDetailed(filename, filenumber)
     {
         _log_level = LogLevel::Warn;
     }
@@ -145,7 +92,8 @@ public:
 
 class LogErrDetailed : public LogDetailed {
 public:
-    LogErrDetailed(const char* filename, int filenumber) : LogDetailed(filename, filenumber)
+    LogErrDetailed(const char* filename, int filenumber) 
+        : LogDetailed(filename, filenumber)
     {
         _log_level = LogLevel::Err;
     }
